Enum futures_await_syn::MacStmtStyle

pub enum MacStmtStyle {
    Semicolon(Semi),
    Braces,
    NoBraces,
}

How a macro was invoked.

Variants
Semicolon(Semi)

The macro statement had a trailing semicolon, e.g. foo! { ... }; foo!(...);, foo![...];

Braces

The macro statement had braces; e.g. foo! { ... }

NoBraces

The macro statement had parentheses or brackets and no semicolon; e.g. foo!(...). All of these will end up being converted into macro expressions.