Enum futures_await_syn::Pat

pub enum Pat {
    Wild(PatWild),
    Ident(PatIdent),
    Struct(PatStruct),
    TupleStruct(PatTupleStruct),
    Path(PatPath),
    Tuple(PatTuple),
    Box(PatBox),
    Ref(PatRef),
    Lit(PatLit),
    Range(PatRange),
    Slice(PatSlice),
    Mac(Mac),
}

Variants
Wild(PatWild)

Represents a wildcard pattern (_)

Ident(PatIdent)

A Pat::Ident may either be a new bound variable (ref mut binding @ OPT_SUBPATTERN), or a unit struct/variant pattern, or a const pattern (in the last two cases the third field must be None). Disambiguation cannot be done with parser alone, so it happens during name resolution.

Struct(PatStruct)

A struct or struct variant pattern, e.g. Variant {x, y, ..}. The bool is true in the presence of a ...

TupleStruct(PatTupleStruct)

A tuple struct/variant pattern Variant(x, y, .., z). If the .. pattern fragment is present, then Option<usize> denotes its position. 0 <= position <= subpats.len()

Path(PatPath)

A possibly qualified path pattern. Unquailfied path patterns A::B::C can legally refer to variants, structs, constants or associated constants. Quailfied path patterns <A>::B::C/<A as Trait>::B::C can only legally refer to associated constants.

Tuple(PatTuple)

A tuple pattern (a, b). If the .. pattern fragment is present, then Option<usize> denotes its position. 0 <= position <= subpats.len()

Box(PatBox)

A box pattern

Ref(PatRef)

A reference pattern, e.g. &mut (a, b)

Lit(PatLit)

A literal

Range(PatRange)

A range pattern, e.g. 1...2

Slice(PatSlice)

[a, b, i.., y, z] is represented as:

Mac(Mac)

A macro pattern; pre-expansion