Enum futures_await_syn::Ty

pub enum Ty {
    Slice(TySlice),
    Array(TyArray),
    Ptr(TyPtr),
    Rptr(TyRptr),
    BareFn(TyBareFn),
    Never(TyNever),
    Tup(TyTup),
    Path(TyPath),
    TraitObject(TyTraitObject),
    ImplTrait(TyImplTrait),
    Paren(TyParen),
    Group(TyGroup),
    Infer(TyInfer),
    Mac(Mac),
}

The different kinds of types recognized by the compiler

Variants
Slice(TySlice)

A variable-length array ([T])

Array(TyArray)

A fixed length array ([T; n])

Ptr(TyPtr)

A raw pointer (*const T or *mut T)

Rptr(TyRptr)

A reference (&'a T or &'a mut T)

BareFn(TyBareFn)

A bare function (e.g. fn(usize) -> bool)

Never(TyNever)

The never type (!)

Tup(TyTup)

A tuple ((A, B, C, D, ...))

Path(TyPath)

A path (module::module::...::Type), optionally "qualified", e.g. <Vec<T> as SomeTrait>::SomeType.

Type parameters are stored in the Path itself

TraitObject(TyTraitObject)

A trait object type Bound1 + Bound2 + Bound3 where Bound is a trait or a lifetime.

ImplTrait(TyImplTrait)

An impl Bound1 + Bound2 + Bound3 type where Bound is a trait or a lifetime.

Paren(TyParen)

No-op; kept solely so that we can pretty-print faithfully

Group(TyGroup)

No-op: kept solely so that we can pretty-print faithfully

Infer(TyInfer)

TyKind::Infer means the type should be inferred instead of it having been specified. This can appear anywhere in a type.

Mac(Mac)

A macro in the type position.

Methods

impl Ty

pub fn without_plus(i: Cursor) -> PResult<Self>