Trait futures_await_syn::fold::Folder

pub trait Folder {
    fn fold_abi(&mut self, i: Abi) -> Abi { ... }

    fn fold_abi_kind(&mut self, i: AbiKind) -> AbiKind { ... }

    fn fold_angle_bracketed_parameter_data(
        &mut self, 
        i: AngleBracketedParameterData
    ) -> AngleBracketedParameterData { ... }

    fn fold_arg_captured(&mut self, i: ArgCaptured) -> ArgCaptured { ... }

    fn fold_arg_self(&mut self, i: ArgSelf) -> ArgSelf { ... }

    fn fold_arg_self_ref(&mut self, i: ArgSelfRef) -> ArgSelfRef { ... }

    fn fold_arm(&mut self, i: Arm) -> Arm { ... }

    fn fold_attr_style(&mut self, i: AttrStyle) -> AttrStyle { ... }

    fn fold_attribute(&mut self, i: Attribute) -> Attribute { ... }

    fn fold_bare_fn_arg(&mut self, i: BareFnArg) -> BareFnArg { ... }

    fn fold_bare_fn_arg_name(&mut self, i: BareFnArgName) -> BareFnArgName { ... }

    fn fold_bare_fn_ty(&mut self, i: BareFnTy) -> BareFnTy { ... }

    fn fold_bin_op(&mut self, i: BinOp) -> BinOp { ... }

    fn fold_binding_mode(&mut self, i: BindingMode) -> BindingMode { ... }

    fn fold_block(&mut self, i: Block) -> Block { ... }

    fn fold_body(&mut self, i: Body) -> Body { ... }

    fn fold_body_enum(&mut self, i: BodyEnum) -> BodyEnum { ... }

    fn fold_body_struct(&mut self, i: BodyStruct) -> BodyStruct { ... }

    fn fold_bound_lifetimes(&mut self, i: BoundLifetimes) -> BoundLifetimes { ... }

    fn fold_capture_by(&mut self, i: CaptureBy) -> CaptureBy { ... }

    fn fold_constness(&mut self, i: Constness) -> Constness { ... }

    fn fold_defaultness(&mut self, i: Defaultness) -> Defaultness { ... }

    fn fold_derive_input(&mut self, i: DeriveInput) -> DeriveInput { ... }

    fn fold_expr(&mut self, i: Expr) -> Expr { ... }

    fn fold_expr_addr_of(&mut self, i: ExprAddrOf) -> ExprAddrOf { ... }

    fn fold_expr_array(&mut self, i: ExprArray) -> ExprArray { ... }

    fn fold_expr_assign(&mut self, i: ExprAssign) -> ExprAssign { ... }

    fn fold_expr_assign_op(&mut self, i: ExprAssignOp) -> ExprAssignOp { ... }

    fn fold_expr_binary(&mut self, i: ExprBinary) -> ExprBinary { ... }

    fn fold_expr_block(&mut self, i: ExprBlock) -> ExprBlock { ... }

    fn fold_expr_box(&mut self, i: ExprBox) -> ExprBox { ... }

    fn fold_expr_break(&mut self, i: ExprBreak) -> ExprBreak { ... }

    fn fold_expr_call(&mut self, i: ExprCall) -> ExprCall { ... }

    fn fold_expr_cast(&mut self, i: ExprCast) -> ExprCast { ... }

    fn fold_expr_catch(&mut self, i: ExprCatch) -> ExprCatch { ... }

    fn fold_expr_closure(&mut self, i: ExprClosure) -> ExprClosure { ... }

    fn fold_expr_continue(&mut self, i: ExprContinue) -> ExprContinue { ... }

    fn fold_expr_field(&mut self, i: ExprField) -> ExprField { ... }

    fn fold_expr_for_loop(&mut self, i: ExprForLoop) -> ExprForLoop { ... }

    fn fold_expr_group(&mut self, i: ExprGroup) -> ExprGroup { ... }

    fn fold_expr_if(&mut self, i: ExprIf) -> ExprIf { ... }

    fn fold_expr_if_let(&mut self, i: ExprIfLet) -> ExprIfLet { ... }

    fn fold_expr_in_place(&mut self, i: ExprInPlace) -> ExprInPlace { ... }

    fn fold_expr_index(&mut self, i: ExprIndex) -> ExprIndex { ... }

    fn fold_expr_kind(&mut self, i: ExprKind) -> ExprKind { ... }

    fn fold_expr_loop(&mut self, i: ExprLoop) -> ExprLoop { ... }

    fn fold_expr_match(&mut self, i: ExprMatch) -> ExprMatch { ... }

    fn fold_expr_method_call(&mut self, i: ExprMethodCall) -> ExprMethodCall { ... }

    fn fold_expr_paren(&mut self, i: ExprParen) -> ExprParen { ... }

    fn fold_expr_path(&mut self, i: ExprPath) -> ExprPath { ... }

    fn fold_expr_range(&mut self, i: ExprRange) -> ExprRange { ... }

    fn fold_expr_repeat(&mut self, i: ExprRepeat) -> ExprRepeat { ... }

    fn fold_expr_ret(&mut self, i: ExprRet) -> ExprRet { ... }

    fn fold_expr_struct(&mut self, i: ExprStruct) -> ExprStruct { ... }

    fn fold_expr_try(&mut self, i: ExprTry) -> ExprTry { ... }

    fn fold_expr_tup(&mut self, i: ExprTup) -> ExprTup { ... }

    fn fold_expr_tup_field(&mut self, i: ExprTupField) -> ExprTupField { ... }

    fn fold_expr_type(&mut self, i: ExprType) -> ExprType { ... }

    fn fold_expr_unary(&mut self, i: ExprUnary) -> ExprUnary { ... }

    fn fold_expr_while(&mut self, i: ExprWhile) -> ExprWhile { ... }

    fn fold_expr_while_let(&mut self, i: ExprWhileLet) -> ExprWhileLet { ... }

    fn fold_expr_yield(&mut self, i: ExprYield) -> ExprYield { ... }

    fn fold_field(&mut self, i: Field) -> Field { ... }

    fn fold_field_pat(&mut self, i: FieldPat) -> FieldPat { ... }

    fn fold_field_value(&mut self, i: FieldValue) -> FieldValue { ... }

    fn fold_file(&mut self, i: File) -> File { ... }

    fn fold_fn_arg(&mut self, i: FnArg) -> FnArg { ... }

    fn fold_fn_decl(&mut self, i: FnDecl) -> FnDecl { ... }

    fn fold_foreign_item(&mut self, i: ForeignItem) -> ForeignItem { ... }

    fn fold_foreign_item_fn(&mut self, i: ForeignItemFn) -> ForeignItemFn { ... }

    fn fold_foreign_item_kind(&mut self, i: ForeignItemKind) -> ForeignItemKind { ... }

    fn fold_foreign_item_static(
        &mut self, 
        i: ForeignItemStatic
    ) -> ForeignItemStatic { ... }

    fn fold_function_ret_ty(&mut self, i: FunctionRetTy) -> FunctionRetTy { ... }

    fn fold_generics(&mut self, i: Generics) -> Generics { ... }

    fn fold_impl_item(&mut self, i: ImplItem) -> ImplItem { ... }

    fn fold_impl_item_const(&mut self, i: ImplItemConst) -> ImplItemConst { ... }

    fn fold_impl_item_kind(&mut self, i: ImplItemKind) -> ImplItemKind { ... }

    fn fold_impl_item_method(&mut self, i: ImplItemMethod) -> ImplItemMethod { ... }

    fn fold_impl_item_type(&mut self, i: ImplItemType) -> ImplItemType { ... }

    fn fold_impl_polarity(&mut self, i: ImplPolarity) -> ImplPolarity { ... }

    fn fold_in_place_kind(&mut self, i: InPlaceKind) -> InPlaceKind { ... }

    fn fold_item(&mut self, i: Item) -> Item { ... }

    fn fold_item_const(&mut self, i: ItemConst) -> ItemConst { ... }

    fn fold_item_default_impl(&mut self, i: ItemDefaultImpl) -> ItemDefaultImpl { ... }

    fn fold_item_enum(&mut self, i: ItemEnum) -> ItemEnum { ... }

    fn fold_item_extern_crate(&mut self, i: ItemExternCrate) -> ItemExternCrate { ... }

    fn fold_item_fn(&mut self, i: ItemFn) -> ItemFn { ... }

    fn fold_item_foreign_mod(&mut self, i: ItemForeignMod) -> ItemForeignMod { ... }

    fn fold_item_impl(&mut self, i: ItemImpl) -> ItemImpl { ... }

    fn fold_item_kind(&mut self, i: ItemKind) -> ItemKind { ... }

    fn fold_item_mod(&mut self, i: ItemMod) -> ItemMod { ... }

    fn fold_item_static(&mut self, i: ItemStatic) -> ItemStatic { ... }

    fn fold_item_struct(&mut self, i: ItemStruct) -> ItemStruct { ... }

    fn fold_item_trait(&mut self, i: ItemTrait) -> ItemTrait { ... }

    fn fold_item_ty(&mut self, i: ItemTy) -> ItemTy { ... }

    fn fold_item_union(&mut self, i: ItemUnion) -> ItemUnion { ... }

    fn fold_item_use(&mut self, i: ItemUse) -> ItemUse { ... }

    fn fold_lifetime_def(&mut self, i: LifetimeDef) -> LifetimeDef { ... }

    fn fold_local(&mut self, i: Local) -> Local { ... }

    fn fold_mac(&mut self, i: Mac) -> Mac { ... }

    fn fold_mac_stmt_style(&mut self, i: MacStmtStyle) -> MacStmtStyle { ... }

    fn fold_meta_item(&mut self, i: MetaItem) -> MetaItem { ... }

    fn fold_meta_item_list(&mut self, i: MetaItemList) -> MetaItemList { ... }

    fn fold_meta_name_value(&mut self, i: MetaNameValue) -> MetaNameValue { ... }

    fn fold_method_sig(&mut self, i: MethodSig) -> MethodSig { ... }

    fn fold_mut_ty(&mut self, i: MutTy) -> MutTy { ... }

    fn fold_mutability(&mut self, i: Mutability) -> Mutability { ... }

    fn fold_nested_meta_item(&mut self, i: NestedMetaItem) -> NestedMetaItem { ... }

    fn fold_parenthesized_parameter_data(
        &mut self, 
        i: ParenthesizedParameterData
    ) -> ParenthesizedParameterData { ... }

    fn fold_pat(&mut self, i: Pat) -> Pat { ... }

    fn fold_pat_box(&mut self, i: PatBox) -> PatBox { ... }

    fn fold_pat_ident(&mut self, i: PatIdent) -> PatIdent { ... }

    fn fold_pat_lit(&mut self, i: PatLit) -> PatLit { ... }

    fn fold_pat_path(&mut self, i: PatPath) -> PatPath { ... }

    fn fold_pat_range(&mut self, i: PatRange) -> PatRange { ... }

    fn fold_pat_ref(&mut self, i: PatRef) -> PatRef { ... }

    fn fold_pat_slice(&mut self, i: PatSlice) -> PatSlice { ... }

    fn fold_pat_struct(&mut self, i: PatStruct) -> PatStruct { ... }

    fn fold_pat_tuple(&mut self, i: PatTuple) -> PatTuple { ... }

    fn fold_pat_tuple_struct(&mut self, i: PatTupleStruct) -> PatTupleStruct { ... }

    fn fold_pat_wild(&mut self, i: PatWild) -> PatWild { ... }

    fn fold_path(&mut self, i: Path) -> Path { ... }

    fn fold_path_glob(&mut self, i: PathGlob) -> PathGlob { ... }

    fn fold_path_list(&mut self, i: PathList) -> PathList { ... }

    fn fold_path_list_item(&mut self, i: PathListItem) -> PathListItem { ... }

    fn fold_path_parameters(&mut self, i: PathParameters) -> PathParameters { ... }

    fn fold_path_segment(&mut self, i: PathSegment) -> PathSegment { ... }

    fn fold_path_simple(&mut self, i: PathSimple) -> PathSimple { ... }

    fn fold_poly_trait_ref(&mut self, i: PolyTraitRef) -> PolyTraitRef { ... }

    fn fold_qself(&mut self, i: QSelf) -> QSelf { ... }

    fn fold_range_limits(&mut self, i: RangeLimits) -> RangeLimits { ... }

    fn fold_stmt(&mut self, i: Stmt) -> Stmt { ... }

    fn fold_trait_bound_modifier(
        &mut self, 
        i: TraitBoundModifier
    ) -> TraitBoundModifier { ... }

    fn fold_trait_item(&mut self, i: TraitItem) -> TraitItem { ... }

    fn fold_trait_item_const(&mut self, i: TraitItemConst) -> TraitItemConst { ... }

    fn fold_trait_item_kind(&mut self, i: TraitItemKind) -> TraitItemKind { ... }

    fn fold_trait_item_method(&mut self, i: TraitItemMethod) -> TraitItemMethod { ... }

    fn fold_trait_item_type(&mut self, i: TraitItemType) -> TraitItemType { ... }

    fn fold_ty(&mut self, i: Ty) -> Ty { ... }

    fn fold_ty_array(&mut self, i: TyArray) -> TyArray { ... }

    fn fold_ty_bare_fn(&mut self, i: TyBareFn) -> TyBareFn { ... }

    fn fold_ty_group(&mut self, i: TyGroup) -> TyGroup { ... }

    fn fold_ty_impl_trait(&mut self, i: TyImplTrait) -> TyImplTrait { ... }

    fn fold_ty_infer(&mut self, i: TyInfer) -> TyInfer { ... }

    fn fold_ty_never(&mut self, i: TyNever) -> TyNever { ... }

    fn fold_ty_param(&mut self, i: TyParam) -> TyParam { ... }

    fn fold_ty_param_bound(&mut self, i: TyParamBound) -> TyParamBound { ... }

    fn fold_ty_paren(&mut self, i: TyParen) -> TyParen { ... }

    fn fold_ty_path(&mut self, i: TyPath) -> TyPath { ... }

    fn fold_ty_ptr(&mut self, i: TyPtr) -> TyPtr { ... }

    fn fold_ty_rptr(&mut self, i: TyRptr) -> TyRptr { ... }

    fn fold_ty_slice(&mut self, i: TySlice) -> TySlice { ... }

    fn fold_ty_trait_object(&mut self, i: TyTraitObject) -> TyTraitObject { ... }

    fn fold_ty_tup(&mut self, i: TyTup) -> TyTup { ... }

    fn fold_type_binding(&mut self, i: TypeBinding) -> TypeBinding { ... }

    fn fold_un_op(&mut self, i: UnOp) -> UnOp { ... }

    fn fold_unsafety(&mut self, i: Unsafety) -> Unsafety { ... }

    fn fold_variant(&mut self, i: Variant) -> Variant { ... }

    fn fold_variant_data(&mut self, i: VariantData) -> VariantData { ... }

    fn fold_view_path(&mut self, i: ViewPath) -> ViewPath { ... }

    fn fold_vis_crate(&mut self, i: VisCrate) -> VisCrate { ... }

    fn fold_vis_inherited(&mut self, i: VisInherited) -> VisInherited { ... }

    fn fold_vis_public(&mut self, i: VisPublic) -> VisPublic { ... }

    fn fold_vis_restricted(&mut self, i: VisRestricted) -> VisRestricted { ... }

    fn fold_visibility(&mut self, i: Visibility) -> Visibility { ... }

    fn fold_where_bound_predicate(
        &mut self, 
        i: WhereBoundPredicate
    ) -> WhereBoundPredicate { ... }

    fn fold_where_clause(&mut self, i: WhereClause) -> WhereClause { ... }

    fn fold_where_eq_predicate(
        &mut self, 
        i: WhereEqPredicate
    ) -> WhereEqPredicate { ... }

    fn fold_where_predicate(&mut self, i: WherePredicate) -> WherePredicate { ... }

    fn fold_where_region_predicate(
        &mut self, 
        i: WhereRegionPredicate
    ) -> WhereRegionPredicate { ... }
}

AST->AST fold.

Each method of the Folder trait is a hook to be potentially overridden. Each method's default implementation recursively visits the substructure of the input via the walk functions, which perform an "identity fold", that is, they return the same structure that they are given (for example the fold_file method by default calls fold::walk_file).

If you want to ensure that your code handles every variant explicitly, you need to override each method. (And you also need to monitor future changes to Folder in case a new method with a new default implementation gets introduced.)

Provided Methods

fn fold_abi(&mut self, i: Abi) -> Abi

fn fold_abi_kind(&mut self, i: AbiKind) -> AbiKind

fn fold_angle_bracketed_parameter_data(
&mut self,
i: AngleBracketedParameterData
) -> AngleBracketedParameterData

fn fold_arg_captured(&mut self, i: ArgCaptured) -> ArgCaptured

fn fold_arg_self(&mut self, i: ArgSelf) -> ArgSelf

fn fold_arg_self_ref(&mut self, i: ArgSelfRef) -> ArgSelfRef

fn fold_arm(&mut self, i: Arm) -> Arm

fn fold_attr_style(&mut self, i: AttrStyle) -> AttrStyle

fn fold_attribute(&mut self, i: Attribute) -> Attribute

fn fold_bare_fn_arg(&mut self, i: BareFnArg) -> BareFnArg

fn fold_bare_fn_arg_name(&mut self, i: BareFnArgName) -> BareFnArgName

fn fold_bare_fn_ty(&mut self, i: BareFnTy) -> BareFnTy

fn fold_bin_op(&mut self, i: BinOp) -> BinOp

fn fold_binding_mode(&mut self, i: BindingMode) -> BindingMode

fn fold_block(&mut self, i: Block) -> Block

fn fold_body(&mut self, i: Body) -> Body

fn fold_body_enum(&mut self, i: BodyEnum) -> BodyEnum

fn fold_body_struct(&mut self, i: BodyStruct) -> BodyStruct

fn fold_bound_lifetimes(&mut self, i: BoundLifetimes) -> BoundLifetimes

fn fold_capture_by(&mut self, i: CaptureBy) -> CaptureBy

fn fold_constness(&mut self, i: Constness) -> Constness

fn fold_defaultness(&mut self, i: Defaultness) -> Defaultness

fn fold_derive_input(&mut self, i: DeriveInput) -> DeriveInput

fn fold_expr(&mut self, i: Expr) -> Expr

fn fold_expr_addr_of(&mut self, i: ExprAddrOf) -> ExprAddrOf

fn fold_expr_array(&mut self, i: ExprArray) -> ExprArray

fn fold_expr_assign(&mut self, i: ExprAssign) -> ExprAssign

fn fold_expr_assign_op(&mut self, i: ExprAssignOp) -> ExprAssignOp

fn fold_expr_binary(&mut self, i: ExprBinary) -> ExprBinary

fn fold_expr_block(&mut self, i: ExprBlock) -> ExprBlock

fn fold_expr_box(&mut self, i: ExprBox) -> ExprBox

fn fold_expr_break(&mut self, i: ExprBreak) -> ExprBreak

fn fold_expr_call(&mut self, i: ExprCall) -> ExprCall

fn fold_expr_cast(&mut self, i: ExprCast) -> ExprCast

fn fold_expr_catch(&mut self, i: ExprCatch) -> ExprCatch

fn fold_expr_closure(&mut self, i: ExprClosure) -> ExprClosure

fn fold_expr_continue(&mut self, i: ExprContinue) -> ExprContinue

fn fold_expr_field(&mut self, i: ExprField) -> ExprField

fn fold_expr_for_loop(&mut self, i: ExprForLoop) -> ExprForLoop

fn fold_expr_group(&mut self, i: ExprGroup) -> ExprGroup

fn fold_expr_if(&mut self, i: ExprIf) -> ExprIf

fn fold_expr_if_let(&mut self, i: ExprIfLet) -> ExprIfLet

fn fold_expr_in_place(&mut self, i: ExprInPlace) -> ExprInPlace

fn fold_expr_index(&mut self, i: ExprIndex) -> ExprIndex

fn fold_expr_kind(&mut self, i: ExprKind) -> ExprKind

fn fold_expr_loop(&mut self, i: ExprLoop) -> ExprLoop

fn fold_expr_match(&mut self, i: ExprMatch) -> ExprMatch

fn fold_expr_method_call(&mut self, i: ExprMethodCall) -> ExprMethodCall

fn fold_expr_paren(&mut self, i: ExprParen) -> ExprParen

fn fold_expr_path(&mut self, i: ExprPath) -> ExprPath

fn fold_expr_range(&mut self, i: ExprRange) -> ExprRange

fn fold_expr_repeat(&mut self, i: ExprRepeat) -> ExprRepeat

fn fold_expr_ret(&mut self, i: ExprRet) -> ExprRet

fn fold_expr_struct(&mut self, i: ExprStruct) -> ExprStruct

fn fold_expr_try(&mut self, i: ExprTry) -> ExprTry

fn fold_expr_tup(&mut self, i: ExprTup) -> ExprTup

fn fold_expr_tup_field(&mut self, i: ExprTupField) -> ExprTupField

fn fold_expr_type(&mut self, i: ExprType) -> ExprType

fn fold_expr_unary(&mut self, i: ExprUnary) -> ExprUnary

fn fold_expr_while(&mut self, i: ExprWhile) -> ExprWhile

fn fold_expr_while_let(&mut self, i: ExprWhileLet) -> ExprWhileLet

fn fold_expr_yield(&mut self, i: ExprYield) -> ExprYield

fn fold_field(&mut self, i: Field) -> Field

fn fold_field_pat(&mut self, i: FieldPat) -> FieldPat

fn fold_field_value(&mut self, i: FieldValue) -> FieldValue

fn fold_file(&mut self, i: File) -> File

fn fold_fn_arg(&mut self, i: FnArg) -> FnArg

fn fold_fn_decl(&mut self, i: FnDecl) -> FnDecl

fn fold_foreign_item(&mut self, i: ForeignItem) -> ForeignItem

fn fold_foreign_item_fn(&mut self, i: ForeignItemFn) -> ForeignItemFn

fn fold_foreign_item_kind(&mut self, i: ForeignItemKind) -> ForeignItemKind

fn fold_foreign_item_static(
&mut self,
i: ForeignItemStatic
) -> ForeignItemStatic

fn fold_function_ret_ty(&mut self, i: FunctionRetTy) -> FunctionRetTy

fn fold_generics(&mut self, i: Generics) -> Generics

fn fold_impl_item(&mut self, i: ImplItem) -> ImplItem

fn fold_impl_item_const(&mut self, i: ImplItemConst) -> ImplItemConst

fn fold_impl_item_kind(&mut self, i: ImplItemKind) -> ImplItemKind

fn fold_impl_item_method(&mut self, i: ImplItemMethod) -> ImplItemMethod

fn fold_impl_item_type(&mut self, i: ImplItemType) -> ImplItemType

fn fold_impl_polarity(&mut self, i: ImplPolarity) -> ImplPolarity

fn fold_in_place_kind(&mut self, i: InPlaceKind) -> InPlaceKind

fn fold_item(&mut self, i: Item) -> Item

fn fold_item_const(&mut self, i: ItemConst) -> ItemConst

fn fold_item_default_impl(&mut self, i: ItemDefaultImpl) -> ItemDefaultImpl

fn fold_item_enum(&mut self, i: ItemEnum) -> ItemEnum

fn fold_item_extern_crate(&mut self, i: ItemExternCrate) -> ItemExternCrate

fn fold_item_fn(&mut self, i: ItemFn) -> ItemFn

fn fold_item_foreign_mod(&mut self, i: ItemForeignMod) -> ItemForeignMod

fn fold_item_impl(&mut self, i: ItemImpl) -> ItemImpl

fn fold_item_kind(&mut self, i: ItemKind) -> ItemKind

fn fold_item_mod(&mut self, i: ItemMod) -> ItemMod

fn fold_item_static(&mut self, i: ItemStatic) -> ItemStatic

fn fold_item_struct(&mut self, i: ItemStruct) -> ItemStruct

fn fold_item_trait(&mut self, i: ItemTrait) -> ItemTrait

fn fold_item_ty(&mut self, i: ItemTy) -> ItemTy

fn fold_item_union(&mut self, i: ItemUnion) -> ItemUnion

fn fold_item_use(&mut self, i: ItemUse) -> ItemUse

fn fold_lifetime_def(&mut self, i: LifetimeDef) -> LifetimeDef

fn fold_local(&mut self, i: Local) -> Local

fn fold_mac(&mut self, i: Mac) -> Mac

fn fold_mac_stmt_style(&mut self, i: MacStmtStyle) -> MacStmtStyle

fn fold_meta_item(&mut self, i: MetaItem) -> MetaItem

fn fold_meta_item_list(&mut self, i: MetaItemList) -> MetaItemList

fn fold_meta_name_value(&mut self, i: MetaNameValue) -> MetaNameValue

fn fold_method_sig(&mut self, i: MethodSig) -> MethodSig

fn fold_mut_ty(&mut self, i: MutTy) -> MutTy

fn fold_mutability(&mut self, i: Mutability) -> Mutability

fn fold_nested_meta_item(&mut self, i: NestedMetaItem) -> NestedMetaItem

fn fold_parenthesized_parameter_data(
&mut self,
i: ParenthesizedParameterData
) -> ParenthesizedParameterData

fn fold_pat(&mut self, i: Pat) -> Pat

fn fold_pat_box(&mut self, i: PatBox) -> PatBox

fn fold_pat_ident(&mut self, i: PatIdent) -> PatIdent

fn fold_pat_lit(&mut self, i: PatLit) -> PatLit

fn fold_pat_path(&mut self, i: PatPath) -> PatPath

fn fold_pat_range(&mut self, i: PatRange) -> PatRange

fn fold_pat_ref(&mut self, i: PatRef) -> PatRef

fn fold_pat_slice(&mut self, i: PatSlice) -> PatSlice

fn fold_pat_struct(&mut self, i: PatStruct) -> PatStruct

fn fold_pat_tuple(&mut self, i: PatTuple) -> PatTuple

fn fold_pat_tuple_struct(&mut self, i: PatTupleStruct) -> PatTupleStruct

fn fold_pat_wild(&mut self, i: PatWild) -> PatWild

fn fold_path(&mut self, i: Path) -> Path

fn fold_path_glob(&mut self, i: PathGlob) -> PathGlob

fn fold_path_list(&mut self, i: PathList) -> PathList

fn fold_path_list_item(&mut self, i: PathListItem) -> PathListItem

fn fold_path_parameters(&mut self, i: PathParameters) -> PathParameters

fn fold_path_segment(&mut self, i: PathSegment) -> PathSegment

fn fold_path_simple(&mut self, i: PathSimple) -> PathSimple

fn fold_poly_trait_ref(&mut self, i: PolyTraitRef) -> PolyTraitRef

fn fold_qself(&mut self, i: QSelf) -> QSelf

fn fold_range_limits(&mut self, i: RangeLimits) -> RangeLimits

fn fold_stmt(&mut self, i: Stmt) -> Stmt

fn fold_trait_bound_modifier(
&mut self,
i: TraitBoundModifier
) -> TraitBoundModifier

fn fold_trait_item(&mut self, i: TraitItem) -> TraitItem

fn fold_trait_item_const(&mut self, i: TraitItemConst) -> TraitItemConst

fn fold_trait_item_kind(&mut self, i: TraitItemKind) -> TraitItemKind

fn fold_trait_item_method(&mut self, i: TraitItemMethod) -> TraitItemMethod

fn fold_trait_item_type(&mut self, i: TraitItemType) -> TraitItemType

fn fold_ty(&mut self, i: Ty) -> Ty

fn fold_ty_array(&mut self, i: TyArray) -> TyArray

fn fold_ty_bare_fn(&mut self, i: TyBareFn) -> TyBareFn

fn fold_ty_group(&mut self, i: TyGroup) -> TyGroup

fn fold_ty_impl_trait(&mut self, i: TyImplTrait) -> TyImplTrait

fn fold_ty_infer(&mut self, i: TyInfer) -> TyInfer

fn fold_ty_never(&mut self, i: TyNever) -> TyNever

fn fold_ty_param(&mut self, i: TyParam) -> TyParam

fn fold_ty_param_bound(&mut self, i: TyParamBound) -> TyParamBound

fn fold_ty_paren(&mut self, i: TyParen) -> TyParen

fn fold_ty_path(&mut self, i: TyPath) -> TyPath

fn fold_ty_ptr(&mut self, i: TyPtr) -> TyPtr

fn fold_ty_rptr(&mut self, i: TyRptr) -> TyRptr

fn fold_ty_slice(&mut self, i: TySlice) -> TySlice

fn fold_ty_trait_object(&mut self, i: TyTraitObject) -> TyTraitObject

fn fold_ty_tup(&mut self, i: TyTup) -> TyTup

fn fold_type_binding(&mut self, i: TypeBinding) -> TypeBinding

fn fold_un_op(&mut self, i: UnOp) -> UnOp

fn fold_unsafety(&mut self, i: Unsafety) -> Unsafety

fn fold_variant(&mut self, i: Variant) -> Variant

fn fold_variant_data(&mut self, i: VariantData) -> VariantData

fn fold_view_path(&mut self, i: ViewPath) -> ViewPath

fn fold_vis_crate(&mut self, i: VisCrate) -> VisCrate

fn fold_vis_inherited(&mut self, i: VisInherited) -> VisInherited

fn fold_vis_public(&mut self, i: VisPublic) -> VisPublic

fn fold_vis_restricted(&mut self, i: VisRestricted) -> VisRestricted

fn fold_visibility(&mut self, i: Visibility) -> Visibility

fn fold_where_bound_predicate(
&mut self,
i: WhereBoundPredicate
) -> WhereBoundPredicate

fn fold_where_clause(&mut self, i: WhereClause) -> WhereClause

fn fold_where_eq_predicate(&mut self, i: WhereEqPredicate) -> WhereEqPredicate

fn fold_where_predicate(&mut self, i: WherePredicate) -> WherePredicate

fn fold_where_region_predicate(
&mut self,
i: WhereRegionPredicate
) -> WhereRegionPredicate