Struct futures_await_syn::PatIdent

pub struct PatIdent {
    pub mode: BindingMode,
    pub ident: Ident,
    pub subpat: Option<Box<Pat>>,
    pub at_token: Option<At>,
}

A Pat::Ident may either be a new bound variable (ref mut binding @ OPT_SUBPATTERN), or a unit struct/variant pattern, or a const pattern (in the last two cases the third field must be None). Disambiguation cannot be done with parser alone, so it happens during name resolution.

Fields
mode: BindingMode ident: Ident subpat: Option<Box<Pat>> at_token: Option<At>