Struct futures_await_syn::QSelf

pub struct QSelf {
    pub lt_token: Lt,
    pub ty: Box<Ty>,
    pub position: usize,
    pub as_token: Option<As>,
    pub gt_token: Gt,
}

The explicit Self type in a "qualified path". The actual path, including the trait and the associated item, is stored separately. position represents the index of the associated item qualified with this Self type.

<Vec<T> as a::b::Trait>::AssociatedItem
 ^~~~~     ~~~~~~~~~~~~~~^
 ty        position = 3

<Vec<T>>::AssociatedItem
 ^~~~~    ^
 ty       position = 0

Fields
lt_token: Lt ty: Box<Ty> position: usize as_token: Option<As> gt_token: Gt