Module futures_await_syn::tokens

Discrete tokens that can be parsed out by synom.

This module contains a number of useful tokens like += and / along with keywords like crate and such. These structures are used to track the spans of these tokens and all implment the ToTokens and Synom traits when the corresponding feature is activated.

Structs

Add

AddEq

And

AndAnd

AndEq

As

At

Bang

Box_

Brace

Bracket

Break

CapSelf

Caret

CaretEq

Catch

Colon

Colon2

Comma

Const

Continue

Crate

Default_

Div

DivEq

Do

Dot

Dot2

Dot3

Else

Enum

Eq

EqEq

Extern

Fn_

For

Ge

Group

Gt

If

Impl

In

LArrow

Le

Let

Loop

Lt

Match

Mod

Move

MulEq

Mut

Ne

Or

OrEq

OrOr

Paren

Pound

Pub

Question

RArrow

Ref

Rem

RemEq

Return

Rocket

Self_

Semi

Shl

ShlEq

Shr

ShrEq

Star

Static

Struct

Sub

SubEq

Super

Trait

Type

Underscore

Union

Unsafe

Use

Where

While

Yield